move fat/ under tools/
- closes #1695
