+++ tools/fat/action_bootstrap.go
