+++ tools/fat/action_miner.go
