+++ tools/fat/action_mining.go
