+++ tools/fat/action_show.go
