+++ tools/fat/environment.go
