+++ tools/fat/environment_test.go
