+++ tools/fat/fatutil/helpers_test.go
