+++ tools/fat/fatutil/stream_recorder.go
