+++ tools/fat/fatutil/stream_recorder_test.go
