+++ tools/fat/genesis.go
