+++ tools/fat/process_stderr_recorder_test.go
